add gcp docs and remove extra (not used) functionality to exec
