+++ docs/development/index.md
@@ -7,7 +7,15 @@ any questions, please [let us know](https://github.com/flux-framework/flux-opera
 ```{toctree}
 :maxdepth: 3
 developer-guide
-aws
-creation
 designs
+debugging
+creation
+```
+
+### Deployment
+
+```{toctree}
+:maxdepth: 3
+aws
+google
 ```
